feature/input-and-button small refactor
