--- src/interfaces/glue-images-area/index.ts
@@ -1,1 +1,0 @@
-export * from './image-layers.config';
