+++ src/shared/svg/closed-eye.icon.tsx
