+++ src/shared/svg/input-cross.icon.tsx
