+++ src/shared/svg/lock.icon.tsx
