+++ src/shared/svg/opened-eye.icon.tsx
