+++ src/shared/svg/upload.icon.tsx
